Broke up the `Query` interface into `Query` and `Submit`
