+++ transactions/demo-typescript/src/lib/index.ts
@@ -16,7 +16,9 @@ import { RtsConfig } from "./rtsconfig.js";
 import { plaPlutusDataToCslPlutusData } from "./utils.js";
 
 import { addVkeyWitness, TxBuilder } from "./tx-builder.js";
-import { Ogmios, Query } from "./query.js";
+import { Query } from "./query.js";
+import { Submit } from "./submit.js";
+import { Ogmios } from "./ogmios.js";
 
 // Patch {@link BigInt} s.t. toJSON can print it for debugging.
 (BigInt.prototype as any).toJSON = function () {
@@ -28,7 +30,7 @@ import { Ogmios, Query } from "./query.js";
  * to the demo protocol
  */
 export class Demo {
-  query: Query;
+  query: Query & Submit;
   #signingKey: csl.PrivateKey;
   #signingKeyAddress: csl.Address;
 
@@ -36,7 +38,7 @@ export class Demo {
    * @internal
    */
   constructor(
-    query: Query,
+    query: Query & Submit,
     signingKey: csl.PrivateKey,
     signingKeyAddress: csl.Address,
   ) {
